image alts, disable useless nextjs lint rule, add suspense to usesearchparams
