+++ src/components/Footer.tsx
@@ -14,9 +14,11 @@ export default function Footer() {
         <a className="underline text-sm" href="mailto:support@stellarcarbon.io">
           support@stellarcarbon.io
         </a>
-        <a href="https://twitter.com/StellarCarbon" target="_blank">
-          <img className="w-6 h-6" src="/x-logo-white.png" />
-        </a>
+        <img
+          className="w-6 h-6"
+          src="/x-logo-white.png"
+          alt="Stellarcarbon Logo"
+        />
       </div>
     </div>
   );
