+++ src/components/PartnerLogo.tsx
@@ -11,7 +11,11 @@ interface PartnerLogoProps {
 export default function PartnerLogo({ partner }: PartnerLogoProps) {
   return (
     <div className="my-8 w-[300px] h-[100px] flex items-center justify-center">
-      <img className="" src={partner} />
+      <img
+        className=""
+        src={partner}
+        alt={`${partner.replace(".png", "")} partner logo`}
+      />
     </div>
   );
 }
